Change Output class to Ui class
