+++ build.gradle
@@ -44,11 +44,13 @@ test {
 }
 
 application {
-    mainClassName = "duke.Duke"
+    //mainClassName = "duke.Duke"
+    mainClassName = "gui.Launcher"
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    //archiveBaseName = "duke"
+    archiveBaseName = "gui"
     archiveClassifier = null
 }
